pyproject.toml
[project]
name = "bananashop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "bcrypt>=4.3.0",
    "fastapi>=0.116.1",
    "google-genai>=1.32.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[email]>=2.11.7",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
    "python-dotenv>=1.1.1",
    "pillow==11.3.0"
]
